Building Data Model: first draft for initial discussion of
The implementation is 90% based on GSMA model.
There are few changes introduced by me.

Also, I am not sure it is worth having the BuildingType, my preference
would be to remove it since we have the “category”. An when introducing
json-ld, category labels will be in a reasoner also supporting
hierarchies.